Add subpackages for base & operations
