--- src/main/java/ca/wolfram/beta/symath/NegateNode.java
@@ -1,38 +1,0 @@
-package ca.wolfram.beta.symath;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-import ca.wolfram.beta.symath.Expression.VMap;
-
-public class NegateNode extends OperationNode{
-
-	private NegateNode(List<MathNode> children) {
-		super(children);
-	}
-	
-    public static MathNode create(List<MathNode> nodes) {
-    	if (nodes.size() > 1) throw new RuntimeException("Negate Node only takes 1 node");
-        return new NegateNode(nodes);
-    }
-
-    public static MathNode create(MathNode... nodes) {
-        return create(new ArrayList<>(Arrays.asList(nodes)));
-    }
-	
-	@Override
-	public NodeType getType() {
-		return NodeType.NEGATE;
-	}
-
-	@Override
-	public double operationEval(VMap variableMap) {
-		return - getChildren().get(0).eval(variableMap);
-	}
-	
-	@Override
-	public String toString() {
-		return String.format("- %s", getChildren().get(0).toString());
-	}
-}
